[Refactor] Shift some methods from projectInputController to AssignmentViewHelper
