+++ docs/DeveloperGuide.adoc
@@ -505,16 +505,16 @@ Step 1) Assume Project has been created and the user is currently managing a spe
 Step 2) User enters the command `assign task -i 1 2 -to 1 2 3 -rm 4`. This indicates the user would like to assign tasks with index
 number 1 and 2 to members 1, 2 and 3, and also unassign/remove the task from member 4.
 
-Step 3) The `ProjectInputController.manageProject()` will call `projectAssignTask()`. A new `AssignmentController` will be created, and the
-user input will be passed into the `assignAndUnassign()` method to manage the assignment.
+Step 3) The `ProjectInputController.manageProject()` will call `projectAssignTask()`. A new `AssignmentController` will be created,
+and the `assignAndUnassign()` method is called to manage the assignment.
 
-Step 4) `AssignmentController` calls `ParserHelper.parseAssignmentParams()` to parse and split the input into 3 parts: the task index numbers, the assignee
+Step 4) `ParserHelper.parseAssignmentParams()` is called to parse and split the input into 3 parts: the task index numbers, the assignee
 index numbers, and the unassignees index numbers.
 
 Step 5) `ParserHelper` checks all 3 sets of index numbers to ensure that they are valid (non-negative integers,
 and exist within the project) using `ParserHelper.parseMembersIndexes()` and `ParserHelper.parseTasksIndexes()`.
 `AssignmentController.checkForSameMemberIndexes()` checks if the list of assignees and unassignees contain any identical
-index numbers, and remove them to avoid redundant work. ParserHelper returns all 3 sets of valid indexes to `AssignmentController`.
+index numbers, and removes them to avoid redundant work. `ParserHelper` returns all 3 sets of valid indexes to `AssignmentController`.
 
 Step 5) If there are valid task numbers, a for loop is used to iterate through the tasks to handle the assignments
 one by one. `AssignmentController.assign()` and `AssignmentController.unassign()` are called to assign or unassign
@@ -525,8 +525,11 @@ avoid any errors (for example, duplicating assignments or trying to remove an as
 The errors are noted down by adding error messages to the ArrayList `errorMessages` which will be displayed to the user later.
 
 Step 7) If the input is valid, the assignment is created by calling `Project.createAssignment()` or removed by calling
-`Project.removeAssignment()`. The HashMaps are manipulated accordingly to note down the assignment between the specified task
-and member. Success messages are stored in `successMessages` and are also displayed to the user later.
+`Project.removeAssignment()`. The 2 HashMaps in the project are manipulated accordingly to note down the assignment between the specified task
+and member. Success messages are stored in `successMessages`.
+
+Step 8) The `errorMessages` and `successMessages` from `AssignmentController` are retrieved by `ProjectInputController`. The messages are validated, and then
+formatted with the help of `ViewHelper`, which organises the information into easy-to-read tables and displayed to the user.
 
 image::images/AssignmentSequenceDiagram.png[]
 
@@ -538,7 +541,6 @@ Step 1) Assume Project has been created and the user is currently managing a spe
 Step 2) User enters a command to view assignments. This can be in one of the following formats:
 
 * `view assignments -m` (to view each member's tasks)
-
 * `view assignments -t` (to view each task's assigned members)
 
 Each command is accompanied with a suffix `all`, or selected index numbers of members or tasks.
@@ -551,16 +553,20 @@ Step 4) Depending on the flag in the user input, `projectViewAssignments()` call
 
 * Member flag `-m`: projectViewMembersAssignments() is called.
 ** `ParserHelper.parseMemberIndexes` is called to parse the valid member indexes.
-** `AssignmentViewHelper.getMemberOutput` is called to retrieve the assignments for the selected members.
-It accesses the HashMap `memberAndIndividualListOfTasks`
-
+** `AssignmentViewHelper.getMemberOutput` is called to access the HashMap `memberAndIndividualListOfTasks`
+to retrieve the assignments for the selected members.
 
 * Task flag `-t`: projectViewTasksAssignments() is called.
 ** `ParserHelper.parseTaskIndexes` is called to parse the valid task indexes.
-** `AssignmentViewHelper.
+** `AssignmentViewHelper.getTaskOutput` is called to access the HashMap to retrieve the assigned members
+for selected tasks.
 
+The necessary information is stored in an ArrayList of String `totalOutputToPrint` in both cases.
 
+Step 5) The helper class `ViewHelper` formats the output in an organised table, and the output is returned
+to `ProjectInputController`, and subsequently back to `CLIView` to be displayed to the user.
 
+image::images/.png[]
 
 ==== Design Considerations
