+++ src/main/java/util/constant/ConstantHelper.java
@@ -55,4 +55,6 @@ public class ConstantHelper {
     public static final String COMMAND_VIEW_ASSIGNMENTS_TASK_FLAG = "-t";
 
     public static final int VALID_VIEW_ASSIGNMENT_LENGTH = 20;
+
+
 }
